+++ RELEASE_NOTES.md
@@ -96,6 +96,8 @@ If you have problems upgrading to `7.0`, first try upgrading to `6.9`. It has a 
 * `StandardKeyboardActionHandler` now requires a `KeyboardController`.
 * `StandardKeyboardActionHandler` `inputViewController` initializer is now a convenience initializer.
 * `StandardKeyboardActionHandler` `changeKeyboardTypeAction` has been removed.
+* `SystemKeyboard` init parameters have been refactored.
+* `SystemKeyboard` `standardKeyboardWidth` has been removed.
 * `SystemKeyboardActionButton` now requires a callout context.
 * `SystemKeyboardButtonRowItem` now requires a callout context.
 * `View` `.keyboardGestures` now requires a callout context.